fix: Prepopulate picture-preview-widget when editing models
Make picture preview widget inherit from ClearableFileInput in order
to show the image source and update js and css to show properly parent
widget. Change widget template to display value in a data attribute
in order to check whether is populated or not and display data attribute
in image source in placeholder in the first case.