fix: Remove conflictive JS methods
Some JS methods cause compatibility problems with Safari: in this case forEach loops
--specially in NodeLists, grabbed with queryDocumentAll() for instance-- and
spread syntax (...array) .
So they've been replaced with vanilla JS solutions. It's remarkable the solution
for the forEach in a node list that has to re replaced with a regular for
iteration with ++i increments, to avoid getting in the iteration the function
item(), that can triggers errors.