feat: Update map controller to add declustering
Update map controller to add declustering features at certain zoom levels.
[see this thread https://github.com/SINTEF-9012/PruneCluster/issues/52]. In
order to do so we enable a listener to moveend events (zoomend is not valid
for all situations) and we change { pruneCluster layer }.Cluster.Size.
This value has to be lesser than 1 to decluster almost completely all the markers
 but it cannot be very low because in that case leaflet markers blink
in an annoying manner.